Add files via upload
This is a Peak Element in an array program. A peak element is the highest element in the array. The peak element is larger than its previous element as well as next element in the array. We can find it using Linear Search method but it will take worst time complexity that is O(n). Hence, we will be using Binary Search method whose time complexity will be O(log n). Note: for the Peak Element, the array must be a Mountain Array.