Add files via upload
Pivot Element in an array is the element which is smaller than its previous element and is greater than its following element. For the pivot element, the array must be sorted(in increasing order) and rotated. Here in this code by using Binary Search method we can find the pivot element in O(log n) time complexity.