Add files via upload
This is the program to find the Square Root of a Square. Here I have used Binary Search to find it. If the user enters an integer element as an input, then it will print its Square Root. It will not print the exact Square Root if entered non-perfect Square. Since Binary Search is used its time complexity is less[O(log n)], compared to Linear Search.