Add files via upload
In this code, an array is an input whose number of elements represent number of books and each element represents number of pages in a book. So this array of books is given and we have to allocate these books to 'm' number of students. We have to allocate the books in such a way that:
1) Each student gets at least one book.
2) Each book should be allocated to a student.
3) Book allocation should be in a contiguous manner.
We have to allocate the books to 'm' students such that the maximum number of pages assigned to a student is minimum.