Add files via upload
I Selection Sort algorithm, we select the smallest(or largest) element from the the unsorted portion of the array and swap it with the first element in the unsorted array. The time complexity of the Selection Sort algorithm is O(n^2) an the Space Complexity for the same is O(1).